Split C++ Func/Table/Memory/Global to their own headers
I did my best to separate out headers but I don't understand enough C++
to fully sever `val.hh` and `func.hh` so they sort of circularly
reference each other now.
